requirements.txt
# FastAPI và dependencies
fastapi
uvicorn[standard]
pydantic
pydantic-settings
python-multipart
aiofiles

# Database
sqlalchemy
psycopg2-binary
pgvector
alembic

# Redis và Queue
redis
rq

# LangChain và AI
langchain
langchain-community
langchain-google-genai
langchain-groq
langchain-text-splitters
google-generativeai
groq
numpy

# Utilities
python-dotenv
beautifulsoup4
wikipedia
requests
rank-bm25
markitdown
aiofiles

# Testing
pytest
httpx
